refactor(codex): split source-family contracts
