+++ crates/ctx-history-capture/src/provider/source_backed/registration/families/jsonl/codex.rs
@@ -6,6 +6,13 @@ use crate::provider::codex::nativepath::{
     ingest_codex_sources_v0, CodexSessionTreeInventoryV0, CodexSourceBackedResultV0,
 };
 
+#[path = "codex_prompt_terminal.rs"]
+mod prompt_terminal;
+use prompt_terminal::{
+    bind_codex_prompt_target, remember_codex_prompt_terminal, revalidate_codex_prompt_inventory,
+    CodexPromptTerminalCapture, CodexPromptTerminalEvidence,
+};
+
 #[cfg(test)]
 type ExplicitCodexStageHook = Box<dyn FnOnce(CodexSourceBackedCountersV0)>;
 
@@ -962,73 +969,6 @@ pub fn register_codex_prompt_history_source_backed_route(
         driver,
     )?);
     Ok(())
-}
-
-type CodexPromptTerminalCapture = dyn Fn(&CertifiedSource) -> Result<CertifiedSourceInventory, SourceBackedRouteError>
-    + Send
-    + Sync;
-
-struct CodexPromptTerminalEvidence {
-    certificate: CertifiedSource,
-    inventory: CertifiedSourceInventory,
-}
-
-fn remember_codex_prompt_terminal(
-    state: &Mutex<Option<CodexPromptTerminalEvidence>>,
-    certificate: CertifiedSource,
-    inventory: CertifiedSourceInventory,
-) -> SourceBackedRouteResult<()> {
-    let mut state = state.lock().map_err(|_| {
-        SourceBackedRouteError::new(
-            SourceBackedRouteErrorKind::Internal,
-            "Codex prompt-history terminal evidence lock was poisoned",
-        )
-    })?;
-    *state = Some(CodexPromptTerminalEvidence {
-        certificate,
-        inventory,
-    });
-    Ok(())
-}
-
-fn bind_codex_prompt_target(
-    state: &Mutex<Option<CodexPromptTerminalEvidence>>,
-    target: SourceBackedRevalidationTarget<'_>,
-) -> bool {
-    let Ok(state) = state.lock() else {
-        return false;
-    };
-    let Some(expected) = state.as_ref() else {
-        return false;
-    };
-    match target {
-        SourceBackedRevalidationTarget::Source(source) => expected.certificate == *source,
-        SourceBackedRevalidationTarget::Deletion(deletion) => {
-            deletion.verifies(&expected.inventory)
-                && !expected
-                    .certificate
-                    .observation()
-                    .source()
-                    .exact_descriptor_eq(deletion.source())
-        }
-    }
-}
-
-fn revalidate_codex_prompt_inventory(
-    state: &Mutex<Option<CodexPromptTerminalEvidence>>,
-    capture: &CodexPromptTerminalCapture,
-    expected_inventory: &CertifiedSourceInventory,
-) -> bool {
-    let Ok(state) = state.lock() else {
-        return false;
-    };
-    let Some(expected) = state.as_ref() else {
-        return false;
-    };
-    if expected.inventory != *expected_inventory {
-        return false;
-    }
-    capture(&expected.certificate).is_ok_and(|inventory| inventory == expected.inventory)
 }
 
 #[cfg(test)]
