+++ crates/ctx-history-capture/tests/active-source-family-contract-tests.txt
@@ -31,7 +31,7 @@ provider::source_backed::family::tests::active_source_family_contract_jsonl_phys
 provider::source_backed::family::tests::active_source_family_contract_jsonl_probe_and_scan_share_one_growing_source
 provider::source_backed::family::tests::active_source_family_contract_jsonl_rewrite_truncate_and_replacement_fail_closed
 provider::source_backed::registration::families::jsonl::codex::tests::active_source_family_contract_prompt_history_terminal_inventory_accepts_deferred_append
-provider::source_backed::tests::codex::active_source_family_contract_explicit_codex_append_catches_up
-provider::source_backed::tests::codex::active_source_family_contract_explicit_codex_defers_append_after_staging
+provider::source_backed::tests::codex_active_contracts::active_source_family_contract_explicit_codex_append_catches_up
+provider::source_backed::tests::codex_active_contracts::active_source_family_contract_explicit_codex_defers_append_after_staging
 provider::sqlite::tests::active_source_family_contract_sqlite_reads_active_wal_without_provider_writes
 provider_sources::sqlite_source::tests::active_source_family_contract_sqlite_keeps_a_pinned_view_and_fails_changed_writer_generation
